generate mermaid code on server
